+++ src/Gaa.Extensions.Observer/IBus.cs
@@ -1,0 +1,33 @@
+namespace Gaa.Extensions;
+
+/// <summary>
+/// Шина для обмена сообщениями и событиями.
+/// </summary>
+public interface IBus
+{
+    /// <summary>
+    /// Публикует сообщение в шину.
+    /// </summary>
+    /// <typeparam name="TMessage">Тип сообщения.</typeparam>
+    /// <param name="message">Сообщение.</param>
+    /// <param name="cancellationToken">Токен отмены.</param>
+    /// <returns>Результат выполнения асинхронной задачи.</returns>
+    Task PublishAsync<TMessage>(
+        TMessage message,
+        CancellationToken cancellationToken)
+        where TMessage : notnull;
+
+    /// <summary>
+    /// Публикует сообщение в шину.
+    /// </summary>
+    /// <typeparam name="TMessage">Тип сообщения.</typeparam>
+    /// <param name="message">Сообщение.</param>
+    /// <param name="messageHeaders">Заголовки сообщения.</param>
+    /// <param name="cancellationToken">Токен отмены.</param>
+    /// <returns>Результат выполнения асинхронной задачи.</returns>
+    Task PublishAsync<TMessage>(
+        TMessage message,
+        IReadOnlyDictionary<string, string> messageHeaders,
+        CancellationToken cancellationToken)
+        where TMessage : notnull;
+}