+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Добавлено
 
+- Добавление проекта `Gaa.Extensions.Observer`.
 - Добавление проекта `Gaa.Extensions.Mediator`.
 
 ### Изменено